+++ example_flutter/lib/main.dart
@@ -1,0 +1,105 @@
+import 'package:example_flutter/user_model.dart';
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  final GlobalKey<FormState> _formKey = GlobalKey();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Form(
+          key: _formKey,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 16,
+                ),
+                child: TextFormField(
+                  decoration: const InputDecoration(
+                    labelText: "Id",
+                  ),
+                  validator: (value) {
+                    return UserModelValidator.validateId(value);
+                  },
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 16,
+                ),
+                child: TextFormField(
+                  decoration: const InputDecoration(
+                    labelText: "Email",
+                  ),
+                  validator: (value) {
+                    return UserModelValidator.validateEmail(value);
+                  },
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 16,
+                ),
+                child: TextFormField(
+                  decoration: const InputDecoration(
+                    labelText: "Password",
+                  ),
+                  validator: (value) {
+                    return UserModelValidator.validatePassword(value);
+                  },
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: validate,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+
+  void validate() {
+    final isValid = _formKey.currentState?.validate() == true;
+    if (!isValid) {
+      return;
+    }
+  }
+}