+++ validator_annotation/test/metadatas_test/numeric_metadata_test/is_zero_validator_test.dart
@@ -54,10 +54,12 @@ void main() {
       description: 'When value is null, should be throw exception',
     );
 
-    testValidatorMetadataFailedValidate(
+    testValidatorMetadataSuccessValidate(
       instance: const IsZeroValidator(),
       description: 'When value is not numeric, should be throw exception',
       valueToValidate: '42',
+      expectedValidateResultValue: false,
+      expectedErrorMessage: 'must be zero',
     );
   });
 }