+++ validator_annotation/test/metadatas_test/numeric_metadata_test/min_value_number_validator_test.dart
@@ -66,12 +66,13 @@ void main() {
       description: '''When value is null, should be throw exception''',
     );
 
-    testValidatorMetadataFailedValidate(
+    testValidatorMetadataSuccessValidate(
       instance: const MinValueNumberValidator(
         value: minValue,
       ),
       description: '''When value is not numeric, should be throw exception''',
       valueToValidate: '42',
+      expectedValidateResultValue: true,
     );
   });
 }