+++ validator_annotation/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0-alpha
 
 - Alpha Version
+
+## 1.0.1-alpha
+
+- Add converter value from string to num for numeric validators
+- Treat exception as an error when validating field and instance